fix: use float instead of integer to worka round max integer issue

+++ packages/services/api/src/modules/schema/module.graphql.ts
@@ -466,7 +466,7 @@ export default gql`
   }
 
   type SchemaCoordinateUsage {
-    total: Int!
+    total: Float!
     isUsed: Boolean!
   }
